backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.11.5
async-timeout==4.0.2
attrs==21.4.0
Authlib==0.15.5
billiard==3.6.4.0
black==22.6.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4
cryptography==37.0.2
dataclasses-json==0.5.7
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
dj-rest-auth==2.2.4
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
Faker==13.12.0
flake8==4.0.1
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.57.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
inotify==0.2.10
isort==5.10.1
jsonschema==4.6.0
kombu==5.2.4
lazy-object-proxy==1.7.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
nose==1.3.7
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
pika==1.2.1
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==4.21.5
psycopg==3.0.14
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pylint==2.13.9
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-inspect==0.7.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
